Fix Clang-Tidy issues

+++ src/fheroes2/editor/editor_interface.cpp
@@ -643,7 +643,7 @@ namespace Interface
                 const int32_t artifactType = _editorPanel.getArtifactType();
 
                 const auto & artifactInfo = Maps::getObjectsByGroup( Maps::ObjectGroup::Artifact );
-                assert( artifactType >= 0 && artifactType < artifactInfo.size() );
+                assert( artifactType >= 0 && artifactType < static_cast<int32_t>( artifactInfo.size() ) );
 
                 // For each Spell Scroll artifact we select a spell.
                 if ( artifactInfo[artifactType].objectType == MP2::OBJ_ARTIFACT && artifactInfo[artifactType].metadata[0] == Artifact::SPELL_SCROLL ) {
@@ -655,7 +655,7 @@ namespace Interface
                     }
 
                     setObjectOnTile( tile, Maps::ObjectGroup::Artifact, _editorPanel.getArtifactType() );
-                    Maps::setSpellOnTile( tile, static_cast<uint32_t>( spellId ) );
+                    Maps::setSpellOnTile( tile, spellId );
                 }
                 else {
                     setObjectOnTile( tile, Maps::ObjectGroup::Artifact, _editorPanel.getArtifactType() );
